Address deadlock diagnostics review feedback
