+++ spec/utils/package-scripts-spec.js
@@ -73,7 +73,6 @@ describe("package scripts", {databaseCleaning: {transaction: true}}, () => {
     await fs.mkdir(temporaryDirectoryParent, {recursive: true})
 
     const temporaryDirectory = await fs.mkdtemp(path.join(temporaryDirectoryParent, "git-install-spec-"))
-    const npmCacheDirectory = path.join(temporaryDirectory, "npm-cache")
     const sourceDirectory = path.join(temporaryDirectory, "source")
     const consumerDirectory = path.join(temporaryDirectory, "consumer")
     const npmExecutable = process.env.npm_execpath
@@ -110,10 +109,10 @@ describe("package scripts", {databaseCleaning: {transaction: true}}, () => {
           "install",
           "--allow-git=all",
           "--allow-remote=all",
-          "--cache", npmCacheDirectory,
           "--ignore-scripts=false",
           "--no-audit",
-          "--no-fund"
+          "--no-fund",
+          "--prefer-offline"
         ],
         cwd: consumerDirectory,
         description: "install Velocious from the temporary Git checkout",
